refactor: one formatRelativeTime func
