+++ ui-poc/src/components/PluginsManagerDialog.tsx
@@ -4,6 +4,7 @@ import { usePluginsStatus } from '../hooks/usePluginOperations';
 import { OperationResource, StatusErrorResource } from '../lib/catalogApi';
 import { PluginStatusBadge } from './PluginStatusBadge';
 import { PluginErrorModal } from './PluginErrorModal';
+import { formatRelativeTime } from '../lib/utils';
 
 interface PluginsManagerDialogProps {
   open: boolean;
@@ -278,20 +279,6 @@ function latestOperationPerPlugin(operations: OperationResource[]): Map<string, 
     }
   }
   return latest;
-}
-
-function formatRelativeTime(iso: string): string {
-  const then = new Date(iso).getTime();
-  const diffMs = Date.now() - then;
-  const seconds = Math.max(0, Math.floor(diffMs / 1000));
-
-  if (seconds < 60) return `just now`;
-  const minutes = Math.floor(seconds / 60);
-  if (minutes < 60) return `${minutes} min ago`;
-  const hours = Math.floor(minutes / 60);
-  if (hours < 24) return `${hours}h ago`;
-  const days = Math.floor(hours / 24);
-  return `${days}d ago`;
 }
 
 /**
